move stat block to view
