+++ js/views/question-third-view.js
@@ -1,4 +1,4 @@
-import {statsBlockElement} from '../elements/stats';
+import StatBlockView from './stat-block-view';
 import {questionsFormElement} from '../elements/questions/form';
 import AbstractView from './abstract-view';
 import {levels, PAINT, PHOTO} from '../data';
@@ -15,12 +15,14 @@ export default class QuestionThirdView extends AbstractView {
   }
 
   get template() {
+    const statBlock = new StatBlockView(this.state.givenAnswers);
+
     return `
       ${this.header.element.innerHTML}
       <div class="game">
         <p class="game__task">${this.questionTitle(levels[this.state.level])}</p>
         ${questionsFormElement(levels[this.state.level])}
-        <div class="stats">${statsBlockElement(this.state.givenAnswers)}</div>
+        <div class="stats">${statBlock.element.innerHTML}</div>
       </div>
       ${this.footer.element.innerHTML}
     `;
