+++ js/views/stat-block-view.js
@@ -1,0 +1,47 @@
+import AbstractView from './abstract-view';
+import {QUESTIONS_COUNT} from '../data';
+import {isFastAnswer, isSlowAnswer} from '../data/points';
+
+const ANSWERS_STATE = {
+  FAST: `fast`,
+  SLOW: `slow`,
+  CORRECT: `correct`,
+  WRONG: `wrong`
+};
+
+export default class StatBlockView extends AbstractView {
+  constructor(answers) {
+    super();
+    this.answers = answers;
+  }
+  get template() {
+    return `
+      <ul class="stats">
+        ${[...this.answers].map((answer) => this.statElement(answer)).join(``)}
+        ${new Array(QUESTIONS_COUNT - this.answers.length)
+          .fill(`<li class="stats__result stats__result--unknown"></li>`)
+          .join(``)}
+      </ul>
+    `;
+  }
+
+  statElement(answer) {
+    return `<li class="stats__result stats__result--${this.statClass(answer)}"></li>`;
+  }
+
+  statClass(answer) {
+    if (!answer.variant) {
+      return ANSWERS_STATE.WRONG;
+    }
+
+    if (isFastAnswer(answer.seconds)) {
+      return ANSWERS_STATE.FAST;
+    }
+
+    if (isSlowAnswer(answer.seconds)) {
+      return ANSWERS_STATE.SLOW;
+    }
+
+    return ANSWERS_STATE.CORRECT;
+  }
+}
